Move ConfigurationImpl to an impl subpackage
