+++ core/src/main/java/org/infinispan/protostream/config/impl/AnnotationAttributeConfigurationImpl.java
@@ -1,4 +1,4 @@
-package org.infinispan.protostream.config;
+package org.infinispan.protostream.config.impl;
 
 import java.util.Arrays;
 import java.util.HashSet;
@@ -6,6 +6,9 @@ import java.util.List;
 import java.util.Set;
 
 import org.infinispan.protostream.AnnotationMetadataCreator;
+import org.infinispan.protostream.config.AnnotationAttributeConfiguration;
+import org.infinispan.protostream.config.AnnotationConfiguration;
+import org.infinispan.protostream.config.Configuration;
 import org.infinispan.protostream.descriptors.AnnotatedDescriptor;
 import org.infinispan.protostream.descriptors.AnnotationElement;
 import org.infinispan.protostream.impl.parser.AnnotationParser;
